Cards for events and expert talks added and news-and-event styled
