+++ next/public/locales/en/common.json
@@ -11,6 +11,7 @@
   "our-work": "Our work",
   "all-works": "All works",
   "our-clients": "Our clients",
+  "our-events": "Our events",
   "coming-events": "Coming events",
   "expert-talks": "Expert talks",
   "promoted-articles": "Promoted articles",
@@ -36,7 +37,7 @@
   "search-go-to-page": "Go to page {{page}}",
   "skip-to-main-content": "Skip to main content",
   "speakers": "Speakers:",
-  "filter-by": "Filter by {{label}}",
+  "filter-by": "Filter by",
   "Page not found": "Page not found",
   "The page you are looking for does not exist.": "The page you are looking for does not exist.",
   "Go back to the homepage?": "Go back to the homepage?",
