+++ next/public/locales/fi/common.json
@@ -11,6 +11,7 @@
   "our-work": "Meidän työmme",
   "all-works": "Kaikki toimii",
   "our-clients": "Asiakkaamme",
+  "our-events": "Tapahtumamme",
   "coming-events": "Tulevat tapahtumat",
   "expert-talks": "Asiantuntijoiden puheenvuorot",
   "promoted-articles": "Suositellut artikkelit",
@@ -36,7 +37,7 @@
   "search-go-to-page": "Mene sivulle {{page}}",
   "skip-to-main-content": "Siirry pääsisältöön",
   "speakers": "Puhujat:",
-  "filter-by": "Suodata {{label}}",
+  "filter-by": "Suodata",
   "Page not found": "Sivua ei löydy",
   "The page you are looking for does not exist.": "Hakemaasi sivua ei löydy.",
   "Go back to the homepage?": "Palaa etusivulle?",
