+++ next/public/locales/sv/common.json
@@ -11,6 +11,7 @@
   "our-work": "Vårt arbete",
   "all-works": "Alla fungerar",
   "our-clients": "Våra kunder",
+  "our-events": "Våra evenemang",
   "coming-events": "Kommande evenemang",
   "expert-talks": "Expertsamtal",
   "promoted-articles": "Rekommenderade artiklar",
@@ -36,7 +37,7 @@
   "search-go-to-page": "Gå till sida {{page}}",
   "skip-to-main-content": "Hoppa till huvudinnehåll",
   "speakers": "Talare:",
-  "filter-by": "Filtrera efter {{label}}",
+  "filter-by": "Filtrera efter",
   "Page not found": "Sidan kunde inte hittas",
   "The page you are looking for does not exist.": "Sidan du söker finns inte.",
   "Go back to the homepage?": "Gå tillbaka till hemsidan?",
